move dependencies in pkgchk
