+++ tools/generator/autorest/generationMetadata.go
@@ -12,8 +12,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/Azure/azure-sdk-for-go/tools/common/packages/track1"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest/model"
-	"github.com/Azure/azure-sdk-for-go/tools/pkgchk/track1"
 )
 
 // GenerationMetadata contains all the metadata that has been used when generating a track 1 package
