+++ tools/generator/cmd/root.go
@@ -15,11 +15,11 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/tools/common/exports"
 	"github.com/Azure/azure-sdk-for-go/tools/common/ioext"
+	"github.com/Azure/azure-sdk-for-go/tools/common/packages/track1"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest/model"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/pipeline"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/utils"
-	"github.com/Azure/azure-sdk-for-go/tools/pkgchk/track1"
 	"github.com/Azure/azure-sdk-for-go/version"
 	"github.com/spf13/cobra"
 )
